Move CharSeqWrapper into LiteralSet to minimize the side affect.
Also revert commit 0387a3b29f57380aed2ecb9d9f6ffd0fa4e302c0 as it is not needed any more.
