+++ api/src/main/java/org/apache/iceberg/expressions/ResidualEvaluator.java
@@ -197,7 +197,7 @@ public class ResidualEvaluator implements Serializable {
     @Override
     public <T> Expression startsWith(BoundReference<T> ref, Literal<T> lit) {
       Comparator<T> cmp = lit.comparator();
-      return  ((String) ref.get(struct)).startsWith(lit.value().toString()) ? alwaysTrue() : alwaysFalse();
+      return  ((String) ref.get(struct)).startsWith((String) lit.value()) ? alwaysTrue() : alwaysFalse();
     }
 
     @Override
